refactor(note): 抽离出note组件
