+++ src/views/admin/Note/index.vue
@@ -1,171 +1,16 @@
 <template>
   <div>
-    <!-- <el-button
-      type="primary"
-      style="margin-bottom: 16px"
-      @click="updateNote"
-      :loading="updating"
-    >
-      更新笔记
-    </el-button> -->
-    <div v-loading="isLoading">
-      <div id="vditor" class="vditor" />
-    </div>
+    <NoteComponent />
   </div>
 </template>
 
 <script>
-import Vditor from "vditor";
-import axios from "axios";
+import NoteComponent from "@/components/Note/NoteComponent.vue";
 
-const defaultText = "##  hello world";
 export default {
   name: "NotePage",
-  data() {
-    return {
-      isLoading: true,
-      isMobile: window.innerWidth <= 960,
-      vditor: null,
-      updating: false,
-    };
-  },
-  created() {
-    this.setDefaultText();
-  },
-  mounted() {
-    this.initVditor();
-    this.$nextTick(() => {
-      this.isLoading = false;
-    });
-  },
-  methods: {
-    initVditor() {
-      const that = this;
-      const options = {
-        theme: "dark",
-        toolbarConfig: {
-          pin: true,
-        },
-        toolbar: [
-          // 你可以根据需要调整按钮顺序和内容
-          {
-            name: "update",
-            tip: "更新笔记",
-            icon: '<i class="el-icon-upload"></i>', // 你可以用 Element 的 icon 或自定义 SVG
-            click() {
-              that.updateNote();
-            },
-          },
-          // 其它 Vditor 默认按钮
-          "emoji",
-          /* "headings", */ "bold",
-          "italic",
-          "strike",
-          "link",
-          "|",
-          "list",
-          "ordered-list",
-          "check",
-          "outdent",
-          "indent",
-          "|",
-          "quote",
-          "line",
-          "code",
-          "inline-code",
-          "insert-before",
-          "insert-after",
-          "|",
-          /* "upload", */ "record",
-          "table",
-          "|",
-          "undo",
-          "redo",
-          "|",
-          "fullscreen",
-          "edit-mode",
-          "more",
-        ],
-        width: this.isMobile ? "100%" : "80%",
-        height: 600,
-        tab: "\t",
-        counter: "999999",
-        typewriterMode: false,
-        mode: "wysiwyg",
-        preview: {
-          delay: 100,
-          show: !this.isMobile,
-        },
-        outline: true,
-        cache: { enable: false },
-        // upload: {
-        //   max: 5 * 1024 * 1024,
-        //   handler(file) {
-        //     let formData = new FormData();
-        //     for (let i in file) {
-        //       formData.append("smfile", file[i]);
-        //     }
-        //     let request = new XMLHttpRequest();
-        //     // request.open("POST", "https://sm.ms/api/upload");
-        //     request.onload = that.onloadCallback;
-        //     request.send(formData);
-        //   },
-        // },
-        after: () => {
-          const savedMdContent =
-            localStorage.getItem("vditorvditor") || defaultText;
-          that.vditor.setValue(savedMdContent);
-        },
-      };
-      this.vditor = new Vditor("vditor", options);
-      this.vditor.focus();
-    },
-    onloadCallback(oEvent) {
-      const currentTarget = oEvent.currentTarget;
-      if (currentTarget.status !== 200) {
-        return this.$message({
-          type: "error",
-          message: currentTarget.status + " " + currentTarget.statusText,
-        });
-      }
-      let resp = JSON.parse(currentTarget.response);
-      let imgMdStr = "";
-      if (resp.code === "invalid_source") {
-        return this.$message({
-          type: "error",
-          message: resp.message,
-        });
-      }
-      if (resp.code === "image_repeated") {
-        imgMdStr = `![](${resp.images})`;
-      } else if (resp.code === "success" || resp.success) {
-        imgMdStr = `![${resp.data.filename}](${resp.data.url})`;
-      }
-      this.vditor.insertValue(imgMdStr);
-    },
-    setDefaultText() {
-      const savedMdContent = localStorage.getItem("vditorvditor") || "";
-      if (!savedMdContent.trim()) {
-        localStorage.setItem("vditorvditor", defaultText);
-      }
-    },
-    async updateNote() {
-      if (!this.vditor) return;
-      this.updating = true;
-      try {
-        // 假设你有 noteId，可以通过路由参数获取
-        const noteId = this.$route.params.note_id;
-        const content = this.vditor.getValue();
-        await axios.put(`http://localhost:8000/notes/${noteId}`, {
-          content,
-        });
-        this.$message.success("笔记已更新");
-      } catch (e) {
-        this.$message.error("更新失败");
-      } finally {
-        this.updating = false;
-      }
-    },
+  components: {
+    NoteComponent,
   },
 };
 </script>
